Support multiple concurrent Flex subscribers
Because the driver manages connection to the Flex device autonomously
(detection and connection) there is currently no need to control the
connection from the client. It therefore makes sense to handle the
serial connection like a resource that the driver constructs or destroys
according to presence of interested subscribes (same model as for RFID
tokens).