Add support for Flex devices on Windows
This makes serial access more robust and portable by switching to the
bug.st/serial library. Windows support is then an easy addition.

In order to use the library, an update to Go 1.12 was required. The
ideal course of action would have been to simply update Nixpkgs.
Unfortunately, anything past 19.03 seems to severely break
cross-building. Various small and non-problematic details would need to
be updated, but in the case of statically-linked musl builds I was not
able to fix the problem after prolonged analysis.

I decided to backport Go 1.12 to Nixpkgs 18.09 instead to unblock
development of the driver. Updating Nixpkgs seems to be tricky enough to
warrant a separate effort.