Remove self-update functionality
The driver is no longer installed standalone in enduser scenarios, but
as part of a OS configuration. It is still useful to publish binaries of
the driver, but they should be retrieved and installed as needed for
research/development needs. The situation where the driver self-updates
in an unsupervise customer installation no longer exists.