Use middleware for CORS headers
Instead of repeatedly setting wildcard CORS headers, draw the
functionality into a simple middleware and restrict to a small number of
permissible origins.

This commit also switches the preflight response to be empty, responding
only with the relevant headers and an empty body.

There is some Internet lore that the appropriate status code for a
successful preflight request is 200[1].

[1]: https://stackoverflow.com/questions/46026409/what-are-proper-status-codes-for-cors-preflight-requests