Make firmware update optimistic
Instead of giving up immediately when we can not detect a Senso in
regular mode or can not send the DFU command, keep going and try to
discover a Senso in bootloader mode that we may be able to transmit the
firmware image to.

This helps us recover from a scenario where a Senso ended up in
bootloader mode but did not receive the firmware due to some network
issue.

Sensos up to at least firmware 3.8.1 do not restart into the regular
mode once brought to bootloader mode, so leaving them there without
means for recovery is problematic.