Provide shell script for darwin cross builds
Building via nix is only supported on MacOS and it's not yet possible
to target a different architecture than the one we're building on.
Working with the system clang and simply cross compiling with go works
smoothly and works across architectures, so it's more sensible to do
that instead. Nix is still being used to provide the build script in
the devshell.