Provide cross-building toolchain in nix shells
By building the binaries via nix derivations we lose the possibility of
using `git describe` to determine the version. We could use `self.shortRev`
or `self.shortDirtyRev` in the flake file, but these are much less informative
than the information provided by `git describe`.

Instead of cross-built derivations, provide development shells with
appropriate tooling, so the output of `git describe` can be passed to the
build process.